code refactorized
